Made pingRoots use PingImage for local files again
